+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.8
 
 - Introduce suggestions as commit message input box quick fix (for VS Code v1.78.0 or newer)
+- Add basic telemetry
 
 ## 0.0.7